+++ libsofia-sip-ua/tport/tport.c
@@ -3172,9 +3172,9 @@ int tport_send_msg(tport_t *self, msg_t *msg,
   for (i = 0, total = 0; i < iovused; i++) {
     if (total + iov[i].mv_len > n) {
       if (tport_is_connection_oriented(self)) {
-	iov[i].mv_len -= n - total;
+	iov[i].mv_len -= (u_long)(n - total);
 	iov[i].mv_base = (char *)iov[i].mv_base + (n - total);
-	if (tport_queue_rest(self, msg, &iov[i], iovused - i) >= 0)
+	if (tport_queue_rest(self, msg, &iov[i], (ssize_t)(iovused - i)) >= 0)
 	  return n;
       }
       else {
@@ -3574,7 +3574,7 @@ void tport_send_queue(tport_t *self)
   size_t i, n, total;
   msg_t *msg;
   msg_iovec_t *iov;
-  int iovused;
+  size_t iovused;
   unsigned short qhead = self->tp_qhead, N = self->tp_params->tpp_qsize;
   su_time_t now;
 
@@ -3600,7 +3600,7 @@ void tport_send_queue(tport_t *self)
     
     for (i = 0, total = 0; i < iovused; i++) {
       if (total + iov[i].mv_len > n) {
-	iov[i].mv_len -= n - total;
+	iov[i].mv_len -= (u_long)(n - total);
 	iov[i].mv_base = (char *)iov[i].mv_base + (n - total);
 
 	self->tp_unsent = iov + i;
